Created new files for Dialogs subcomponents
