+++ src/components/Dialogs/Dialogs.jsx
@@ -1,19 +1,8 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
 import style from "./Dialogs.module.css";
-
-const DialogItem = (props) => {
-  let path = "/dialogs/" + props.id;
-  return (
-    <div className={style.dialog}>
-      <NavLink to={path}>{props.name}</NavLink>
-    </div>
-  );
-};
-
-const Message = (props) => {
-  return <div className={style.message}>{props.message}</div>;
-};
+import DialogItem from "./DialogsItem/DialogItem";
+import Message from "./Message/Message";
 
 const Dialogs = (props) => {
   let dialogsData = [
