+++ src/components/Dialogs/Dialogs.module.css
@@ -4,7 +4,7 @@
 }
 
 .dialogsItems .active {
-  color: brown;
+  color: rgb(255, 255, 255);
 }
 
 .messages {
