+++ src/components/Dialogs/Message/Message.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+import { NavLink } from "react-router-dom";
+import style from "./../Dialogs.module.css";
+
+const Message = (props) => {
+  return <div className={style.message}>{props.message}</div>;
+};
+
+export default Message;
